Add files via upload
Looking at case studies from around the world, women are the most exploited group as laborers and a highly tapped into demographic of consumers. I will use the global apparel industry as an example of capital consolidation and worker exploitation and the beauty industry in Taiwan as the generation of a new realm of capital and desire.