Add files via upload
This paper will examine the history of homosexuality and paraphilias, compare the categories against the criteria for a mental disorder, and highlight controversies in disordering alternate sexualities. The author will argue that there is no way to justify the inclusion of lawful paraphilic disorders as a category in the DSM-5, or future DSMs. The inclusion of non-lawful paraphilic disorders can also be contested, but under slightly different considerations.