Add files via upload
Andaz is a film with strong moralistic undertones. It includes many tropes and techniques that we have seen in other social films such as Pyaasa, Awara, and Mother India. It includes themes of family conflict, dissolution of social identity and melodrama (“Shifting Codes”).  In this essay, I will examine scenes and filmic elements that contribute to coding this movie as a social film.